requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data handling
h5py>=3.6.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Configuration
pyyaml>=6.0

# Progress bars and logging
tqdm>=4.62.0

# Jupyter support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Visualization
plotly>=5.0.0

# Optional: TensorBoard for training visualization
tensorboard>=2.10.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.13.0

# Optional: Hyperparameter optimization
# optuna>=3.0.0

# Utils
scikit-image>=0.19.0
